hw_4/requirements.txt
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
h11==0.14.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
pycparser==2.22
PySocks==1.7.1
pytest==8.2.2
pytest-html==4.1.1
pytest-html-reporter==0.2.9
pytest-metadata==3.1.1
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.1
urllib3==2.2.1
webdriver-manager==4.0.1
webdrivermanager-cn==1.7.1
wsproto==1.2.0
